re organisaed the code a little
The UI is generated and I separated the Input into a sperate script to allow more keys to be added.
