+++ StudioUnityProject/Assets/Prefabs/UI.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 9fc0d4010bbf28b4594072e72b8655ab
+guid: 40446d8ce3958f64d9a32a7beeb62006
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
