+++ StudioUnityProject/Assets/Prefabs/UI/BetterInputSystem.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 114d6deb3b86f6e4dbe66baac2df196d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
