+++ StudioUnityProject/Assets/Scripts/BetterInputSystem/InputActions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9941378b517e06d4586a2e5b197796c5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
