+++ StudioUnityProject/Assets/Scripts/UI/BetterInputUI/BetterInputUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b774509b9ab546142a5379c5acd039a4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
